requirements.txt
openai-agents
openai
pandas
duckdb
pydantic
pyarrow
python-dotenv